+++ modules/common/ghostship-pkg.nix
@@ -1,6 +1,5 @@
 self: super: {
   agent-deck = super.callPackage ../../pkgs/agent-deck.nix { };
-  workmux = super.callPackage ../../pkgs/workmux.nix { };
 
   ghostship-config = super.writers.writePython3Bin "ghostship-config" {
     libraries = with super.python3Packages; [