+++ pkgs/agent-deck.nix
@@ -2,13 +2,13 @@
 
 buildGoModule rec {
   pname = "agent-deck";
-  version = "1.3.4";
+  version = "1.4.1";
 
   src = fetchFromGitHub {
     owner = "asheshgoplani";
     repo = "agent-deck";
     tag = "v${version}";
-    hash = "sha256-x4PRqFZJ102+QQzOlIxZbQo/fATK12/Et27A1WflkZs=";
+    hash = "sha256-7upPKCsuKVSMqyimFECVTQ1kJYhk73F6oDKuowejzLQ=";
   };
 
   vendorHash = "sha256-qKK9Wu5+0bi+x6/OwRueIvPi6f4hFUqG+RkhWnLOr5Q=";